pyproject.toml
[tool.poetry]
name = "net-watcher"
version = "0.1.0"
description = ""
authors = ["Artem Kotik <miaow2@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-filter = "^2.4.0"
requests = "^2.24.0"
nornir = "^3.0.0"
hvac = "^0.10.5"
nornir-scrapli = "^2020.11.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
